requirements.txt
azure-functions
requests-toolbelt
